EntityGraph refactoring

+++ src/main/engine/net/sf/jailer/entitygraph/local/LocalEntityGraph.java
@@ -392,8 +392,8 @@ public class LocalEntityGraph extends EntityGraph {
 	 * Deletes the graph.
 	 */
 	@Override
-	public void delete() throws SQLException {
-		if (!isTruncated) {
+	public void delete(boolean forced) throws SQLException {
+		if (!isTruncated && forced) {
 			localSession.executeUpdate("Delete from " + dmlTableReference(DEPENDENCY, localSession) + " Where r_entitygraph=" + graphID + "");
 			localSession.executeUpdate("Delete from " + dmlTableReference(ENTITY, localSession) + " Where r_entitygraph=" + graphID + "");
 			localSession.executeUpdate("Delete from " + dmlTableReference(ENTITY_GRAPH, localSession) + " Where id=" + graphID + "");
@@ -884,31 +884,7 @@ public class LocalEntityGraph extends EntityGraph {
 			}
 		});
 	}
-	
-	/**
-	 * Unites the graph with another one and deletes the other graph.
-	 * 
-	 * @param graph the graph to be united with this graph
-	 */
-	@Override
-	public void uniteWith(EntityGraph graph) throws SQLException {
-		StringBuffer e1EqualsE2 = new StringBuffer();
-		for (Column column: universalPrimaryKey.getColumns()) {
-			if (e1EqualsE2.length() > 0) {
-				e1EqualsE2.append(" and ");
-			}
-			e1EqualsE2.append("E1." + column.name + "=E2." + column.name);
-		}
-		localSession.executeUpdate("Update " + dmlTableReference(ENTITY, localSession) + " E1 " +
-				"set E1.r_entitygraph=" + graphID + " " +
-				"Where E1.r_entitygraph=" + graph.graphID + " " +
-				"and not exists(Select * from " + dmlTableReference(ENTITY, localSession) + " E2 Where " +
-				"E2.r_entitygraph=" + graphID + " and " +
-				e1EqualsE2 +
-				")");
-		graph.delete();
-	}
-	
+
 	/**
 	 * Reads all entities of a given table.
 	 *
